chore: v1 assumes environments are pre-installed
Remove v1's on-demand install from the Environments Hub. `ensure_installed`
auto-pulled a hub `org/name[@version]` env via `install_from_hub` during
plugin/taskset loading; v1 now assumes the package is already importable and
resolves it by module name only. The platform (prime CLI + platform) is
responsible for installing environments.

- Drop `ensure_installed`; rename `v1/utils/install.py` -> `ids.py` (now pure
  id helpers, matching ARCHITECTURE.md's reference).
- `loaders.py` / `legacy.py` resolve via `env_module` instead of installing.
- v0 (`vf-install`, `install_utils`, the prime plugin, `vf-eval` gating) is
  left untouched.
