+++ verifiers/v1/legacy.py
@@ -290,12 +290,11 @@ class LegacyEnvServer(EnvServer):
         extra_env_kwargs: dict | None = None,
     ) -> None:
         from verifiers import load_environment
-        from verifiers.v1.utils.install import ensure_installed, env_name
+        from verifiers.v1.utils.ids import env_module, env_name
 
         self.address = address
-        # Install from the env hub on demand for an `org/name[@version]` id, then load the
-        # v0 env by its module name (a local id is already importable).
-        module = ensure_installed(env_id)
+        # Load the v0 env by its module name; the package must already be installed.
+        module = env_module(env_id)
         self.taskset_id = env_name(env_id)
         self.env = load_environment(module, **(env_args or {}))
         if extra_env_kwargs:  # post-load knobs applied via the v0 env's setters
@@ -421,7 +420,7 @@ def _eval_client(client_config: ClientConfig, model: str):
 def _legacy_output_dir(config) -> Path:
     """The legacy run's output dir, mirroring the native `output_path` shape but keyed by
     the v0 env id (`outputs/<id>--<model>--legacy/<uuid>`); honors `--output-dir`."""
-    from verifiers.v1.utils.install import env_name
+    from verifiers.v1.utils.ids import env_name
 
     if config.output_dir is not None:
         return config.output_dir
@@ -445,11 +444,10 @@ async def run_legacy_eval(config) -> list[Trace]:
     from verifiers import load_environment
 
     from verifiers.v1.cli.output import append_trace, save_config
-    from verifiers.v1.utils.install import ensure_installed
+    from verifiers.v1.utils.ids import env_module
 
-    # Install from the env hub on demand for an `org/name[@version]` id (a local id is
-    # already importable), then load by module name.
-    env = load_environment(ensure_installed(config.id), **(config.args or {}))
+    # Load the v0 env by its module name; the package must already be installed.
+    env = load_environment(env_module(config.id), **(config.args or {}))
     if config.extra_env_kwargs:  # post-load knobs (max_total_completion_tokens, …)
         env.set_kwargs(**config.extra_env_kwargs)
     dataset = env.get_eval_dataset()  # the eval split (falls back to train when unset)
